fix: complete VS-V2 C++ libs build and candle-engine tests
- Fix capital-client curl/REST adapter (CapitalClient, renamed headers)
- Fix candle builder close-once semantics and deduplicator edge cases
- Wire market-core as static lib; all 24 libs compile under CMake
- Update tests for new architecture; 9/9 ctest targets pass

+++ libs/market-types/include/mr/market_types/timeframe.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 namespace mr {
-enum class Timeframe : std::uint8_t { Second1=1, Second10=10, Minute1=60, Minute5=300, Minute15=900, Hour1=3600 };
+enum class Timeframe : std::uint32_t { Second1=1, Second10=10, Minute1=60, Minute5=300, Minute15=900, Hour1=3600 };
 inline std::uint64_t timeframe_ns(Timeframe tf) {
     switch(tf) {
         case Timeframe::Second1: return 1'000'000'000ULL;
